AuthCommander 클래스를 역할에 맞게 분리
