+++ src/main/java/nextstep/auth/controller/TokenController.java
@@ -4,7 +4,8 @@ import lombok.RequiredArgsConstructor;
 import nextstep.auth.controller.dto.OAuthCodeRequest;
 import nextstep.auth.controller.dto.TokenRequest;
 import nextstep.auth.controller.dto.TokenResponse;
-import nextstep.auth.domain.command.AuthCommander;
+import nextstep.auth.domain.command.SocialOAuthUserAuthenticator;
+import nextstep.auth.domain.command.EmailPasswordAuthenticator;
 import nextstep.auth.domain.entity.SocialOAuthProvider;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,17 +15,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 public class TokenController {
-    private final AuthCommander authCommander;
+    private final EmailPasswordAuthenticator emailPasswordAuthenticator;
+    private final SocialOAuthUserAuthenticator socialOAuthUserAuthenticator;
 
     @PostMapping("/login/token")
     public ResponseEntity<TokenResponse> createToken(@RequestBody TokenRequest request) {
-        String token = authCommander.authenticateEmailPassword(request.getEmail(), request.getPassword());
+        String token = emailPasswordAuthenticator.authenticate(request.getEmail(), request.getPassword());
         return ResponseEntity.ok(new TokenResponse(token));
     }
 
     @PostMapping("/login/github")
     public ResponseEntity<TokenResponse> createToken(@RequestBody OAuthCodeRequest request) {
-        String token = authCommander.authenticateSocialOAuth(request.toCommand(SocialOAuthProvider.GITHUB));
+        String token = socialOAuthUserAuthenticator.authenticate(request.toCommand(SocialOAuthProvider.GITHUB));
         return ResponseEntity.ok(new TokenResponse(token));
     }
 }
