+++ src/main/java/nextstep/auth/domain/command/SocialOAuthUserAuthenticator.java
@@ -7,18 +7,12 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class AuthCommander {
-    private final EmailPasswordVerifier emailPasswordVerifier;
+public class SocialOAuthUserAuthenticator {
     private final SocialOAuthUserVerifier socialOAuthUserVerifier;
     private final SocialOAuthUserFetcher socialOAuthUserFetcher;
     private final TokenGenerator tokenGenerator;
 
-    public String authenticateEmailPassword(String email, String password) {
-        TokenPrincipal tokenPrincipal = emailPasswordVerifier.verify(email, password);
-        return tokenGenerator.createToken(tokenPrincipal);
-    }
-
-    public String authenticateSocialOAuth(AuthenticateSocialOAuthCommand.ByAuthCode command) {
+    public String authenticate(AuthenticateSocialOAuthCommand.ByAuthCode command) {
         SocialOAuthUser socialOAuthUser = socialOAuthUserFetcher.fetch(command);
         TokenPrincipal tokenPrincipal = socialOAuthUserVerifier.verify(socialOAuthUser);
         return tokenGenerator.createToken(tokenPrincipal);
